Small Patch V2.1.1
V2.1.1:
Mainly cleaned out cardRenderer.js and utils.js

Will do a new card type maybe later but nrn since i dont got that much time in my grubby hands.
