+++ middleware/renderRankGraph.js
@@ -1,5 +1,3 @@
-const { log } = require("./utils.js");
-
 /**
  * Renders a SVG graph of a player's rank history like in the osu website
  * @param {object} rankHistory - The rank history object
