+++ middleware/utils/log.js
@@ -1,0 +1,11 @@
+/**
+ * Logs a silly message (Just using it cuz smol)
+ * @param {string} string - The message to log
+ */
+const log = (string) => {
+    console.log(string);
+};
+
+module.exports = {
+    log,
+};
