+++ package.json
@@ -1,6 +1,6 @@
 {
     "name": "osu-profile-stats",
-    "version": "2.1.0",
+    "version": "2.1.1",
     "main": "server.js",
     "scripts": {
         "test": "node testServer",
