+++ routes/userData.js
@@ -2,8 +2,8 @@ const express = require("express");
 const router = express.Router();
 const { log } = require("../middleware/utils.js");
 const { dateTan } = require("datetan");
-const { getOsuToken, fetchUserData } = require("../services/osuApi");
-const { metrics } = require("../services/metrics");
+const { getOsuToken, fetchUserData } = require("../services/osuApi.js");
+const { metrics } = require("../services/metrics.js");
 
 router.get("/:username", async (req, res) => {
     try {
